requirements.txt
fastapi
uvicorn
pytest
requests
pyodbc
python-dotenv